// WADNR.Web/src/app/shared/generated/model/organization-type-upsert-request.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class OrganizationTypeUpsertRequest { 
    OrganizationTypeName?: string | null;
    OrganizationTypeAbbreviation?: string | null;
    LegendColor?: string | null;
    ShowOnProjectMaps?: boolean;
    IsDefaultOrganizationType?: boolean;
    IsFundingType?: boolean;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface OrganizationTypeUpsertRequestForm { 
    OrganizationTypeName?: FormControl<string>;
    OrganizationTypeAbbreviation?: FormControl<string>;
    LegendColor?: FormControl<string>;
    ShowOnProjectMaps?: FormControl<boolean>;
    IsDefaultOrganizationType?: FormControl<boolean>;
    IsFundingType?: FormControl<boolean>;
}

export class OrganizationTypeUpsertRequestFormControls { 
    public static OrganizationTypeName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static OrganizationTypeAbbreviation = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static LegendColor = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ShowOnProjectMaps = (value: FormControlState<boolean> | boolean = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<boolean>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static IsDefaultOrganizationType = (value: FormControlState<boolean> | boolean = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<boolean>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static IsFundingType = (value: FormControlState<boolean> | boolean = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<boolean>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}